skins menu preview
ok I'll explaing my thinking: to make electron preview I would make electron sprites with activated skins on the buttons. but to keep the animation and trails I'd have to keep them as game objects. but keeping them as game objects would hide them behind the ui panels.so Im trying to figure out how to get the gameobject to go over the ui, and failing to do that.